backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
aioredis==2.0.1
redis==5.0.1

# Celery & Task Queue
celery==5.3.4
redis==5.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
clerk-sdk-python==0.2.0

# Rate Limiting
slowapi==0.1.8

# Structured Logging
structlog==23.2.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# ML/Data Science
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Configuration
pydantic-settings==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
factory-boy==3.3.0
freezegun==1.2.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
ruff==0.1.6